+++ src/core/include/mp-units/utility/safe_int.h
@@ -27,7 +27,7 @@
 #include <mp-units/bits/module_macros.h>
 #include <mp-units/utility/constrained.h>
 #if MP_UNITS_HOSTED
-#include <mp-units/bits/fmt.h>
+#include <mp-units/ext/format.h>
 #endif
 #include <mp-units/framework/customization_points.h>
 #include <mp-units/framework/representation_concepts.h>